fix(onboard): return durable Hermes registration
